apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
fastapi = {version = ">=0.115.12", python = ">=3.13,<4.0"}
langchain = {version = ">=0.3.23", python = ">=3.13,<4.0"}
langchain-community = {version = ">=0.3.21", python = ">=3.13,<4.0"}
langchain-openai = {version = ">=0.3.12", python = ">=3.13,<4.0"}
langchain-text-splitters = {version = ">=0.3.8", python = ">=3.13,<4.0"}
mypy = {version = ">=1.15.0", python = ">=3.13,<4.0"}
pydantic = {version = ">=2.11.2", python = ">=3.13,<4.0"}
python-multipart = {version = ">=0.0.20", python = ">=3.13,<4.0"}
uvicorn = {version = ">=0.34.0", python = ">=3.13,<4.0"}
unstructured-client = {version = ">=0.32.1,<0.33.0", python = ">=3.13,<4.0"}
unstructured = {version = ">=0.17.2,<0.18.0", python = ">=3.13,<4.0"}
pypandoc = {version = ">=1.15,<2.0", python = ">=3.13,<4.0"}
langchain-weaviate = {version = ">=0.0.4,<0.0.5", python = ">=3.13,<4.0"}
weaviate-client = {version = ">=4.12.0,<5.0.0", python = ">=3.13,<4.0"}
sqlalchemy = {version = ">=2.0.40,<3.0.0", python = ">=3.13,<4.0"}
psycopg2 = {version = ">=2.9.10,<3.0.0", python = ">=3.13,<4.0"}
google-cloud-storage = {version = ">=3.1.0,<4.0.0", python = ">=3.13,<4.0"}
pyhumps = "^3.8.0"
google-cloud-texttospeech = "^2.27.0"
google-generativeai = "^0.8.5"
pydub = "^0.25.1"
ebooklib = "^0.19"
audioop-lts = "^0.2.1"
huggingface-hub = "^0.26.0"
sentence-transformers = "^3.3.1"
langchain-huggingface = "^0.1.2"
langchain-ollama = "^0.2.0"

[tool.mypy]
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]
ignore_missing_imports = true
strict_optional = true
check_untyped_defs = true
strict_equality = true
warn_unused_ignores = true
warn_redundant_casts = true


[tool.pydantic-mypy]
init_forbid_extra = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.poetry.group.dev.dependencies]
pre-commit = ">=4.0.1,<5.0.0"
